Fix enchantment table interaction
All interaction handling for enchantment tables was moved to the server. This is due to a race condition between Vanilla interaction packets and the previous method of handling the UI. As a consequence, ray tracing is slightly inaccurate, but if a player stands still in front of the table, this should not be a problem.