Add basic interaction to beacon and enhance visuals
Visually, the beacon now shows whether the active effect is of level 2 and changes its appearance when in edit mode. The mode is activated by right clicking the block with an empty hand, the effect can be changed by scrolling. In contrast to Vanilla mechanics, the beacon does not require payment in order to activate effects.